Support for absolute/relative paths with .. while uploading sources (#209)
Support for upload source files with paths like:
- absolute paths (Unix style / or C:\ for Windows)
- relative paths outside of current directory (..)

These feature is available for Python version 3.5 or later. If enabled, common path of all found files will be calculated and paths presented on server will be relative to calculated common path. To be backwards compatible, common path is guaranteed not to be subpath of current directory.

For Python version earlier then 3.5, behaviour does not change  - paths presented on server are relative to user's current directory and paths outside of current directory are not allowed.