Add a separate exception for VISIBILITY_RESTRICTED
This error occurs when create_project() calls for a visibility setting
that is not available to the user (for example: using "priv" setting in
a workspace with "Team" plan).
This error uses HTTP code 422 (unprocessable entity), which
neptune-client reported as "(project number) limit reached". That's
because this error uses the same HTTP code, and recognizing the error
based on "errorType" field in the body had to be introduced.